Bumped GWT to 2.7.0.
Also disabled GWT htmlunit tests because it seems to choke on the fact that I
factored the GWT bits into src/main/gwt.

It should work fine because everything is in target/classes by the time GWT
runs, and target/classes is on GWT's classpath, but it doesn't. I'm sure
someone is doing something clever along the way and passing the value on to the
customer.